+++ src/fem_domain/basis.rs
@@ -8,7 +8,6 @@ use super::domain::mesh::{
 use crate::matrix_math::integration::glq::{gauss_quadrature_points, scale_gauss_quad_points};
 
 use std::collections::HashMap;
-use std::rc::Rc;
 use std::sync::{Arc, Mutex};
 
 /// A 1D Higher-Order Hierarchical Shape Function Space: defined on (-1.0, +1.0)
@@ -47,6 +46,100 @@ pub trait ShapeFn: Clone + Sync + Send + std::fmt::Debug {
     fn norm_d2(&self, n: usize, p: usize) -> f64;
 }
 
+// /// A utility structure used to generate and cache [BasisFn]'s during Integration
+// ///
+// /// Three important areas of functionality are handled here:
+// /// * Convert `BasisSpec`s (abstract basis function definitions on an [Elem]) into [BasisFn]'s (numerical instantiations of basis functions for integration)
+// /// * Handle interpolation between parametric spaces to set up inter-layer integration
+// /// * Cache computed [BasisFn]'s to avoid re-computation.
+// ///
+// #[derive(Debug)]
+// pub struct BasisFnSampler<SF: ShapeFn> {
+//     /// Maximum u-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`
+//     pub i_max: usize,
+//     /// Maximum v-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`
+//     pub j_max: usize,
+//     /// Whether the 2nd derivatives of the [ShapeFn]s will be computed. If true, the endpoints [-1, +1] are also defined along both axes!
+//     pub compute_d2: bool,
+//     /// Gauss Legendre Quadrature points evaluated along u-direction. Defined from (-1 to +1)
+//     u_points: Vec<f64>,
+//     /// Gauss Legendre Quadrature points evaluated along v-direction. Defined from (-1 to +1)
+//     v_points: Vec<f64>,
+
+//     computed: HashMap<BSDescription, Rc<BasisFn<SF>>>,
+// }
+
+// impl<SF: ShapeFn> BasisFnSampler<SF> {
+//     /// Construct a sampler with the following parameters:
+//     ///
+//     /// * `i_max` : maximum u-directed expansion order
+//     /// * `j_max` : maximum v-directed expansion order
+//     /// * `num_u_points` : number of Gauss-Leg-Quad points to define along the u-axis. If `None`: a default value is used
+//     /// * `num_v_points` : number of Gauss-Leg-Quad points to define along the v-axis. If `None`: a default value is used
+//     /// * `compute_2nd_derivs`: whether or not to compute the second derivatives of the [ShapeFn]s.
+//     ///
+//     /// A [ShapeFn] implementation must also be specified.
+//     ///
+//     /// These parameters will be used to construct all [BasisFn]s generated by this sampler.
+//     ///
+//     /// The default values for `num_u_points` and `num_v_points` are cautiously large. Smaller values will result in much faster computation, but may result in a loss of accuracy depending on the implementation of [BasisFn]. For `KOLBasis`, n+2 is sufficient for accurate integration.
+//     pub fn with(
+//         i_max: usize,
+//         j_max: usize,
+//         num_u_points: Option<usize>,
+//         num_v_points: Option<usize>,
+//         compute_2nd_derivs: bool,
+//     ) -> (Self, [Vec<f64>; 2]) {
+//         let (u_points, u_weights) = gauss_quadrature_points(
+//             num_u_points.unwrap_or_else(|| default_ngq(i_max)),
+//             compute_2nd_derivs,
+//         );
+//         let (v_points, v_weights) = gauss_quadrature_points(
+//             num_v_points.unwrap_or_else(|| default_ngq(j_max)),
+//             compute_2nd_derivs,
+//         );
+
+//         (
+//             Self {
+//                 i_max,
+//                 j_max,
+//                 compute_d2: compute_2nd_derivs,
+//                 u_points,
+//                 v_points,
+//                 computed: HashMap::new(),
+//             },
+//             [u_weights, v_weights],
+//         )
+//     }
+
+//     /// Generate or retrieve a [BasisFn] defined over an [Elem]. Can be defined over a subset of the Element.
+//     ///
+//     /// If a descendant `Elem` is provided the [BasisFn] will be sampled over the region of the `Elem` that is overlapped by the descendant. Otherwise, the [BasisFn] will be sampled over the entire `Elem`.
+//     pub fn sample_basis_fn(
+//         &mut self,
+//         elem: &Elem,
+//         over_desc_elem: Option<&Elem>,
+//     ) -> Rc<BasisFn<SF>> {
+//         let desc = BSDescription::new(elem, over_desc_elem);
+
+//         if let Some(computed_bs) = self.computed.get(&desc) {
+//             computed_bs.clone()
+//         } else {
+//             let bs = BasisFn::mapped_over_desc(
+//                 self.i_max,
+//                 self.j_max,
+//                 self.compute_d2,
+//                 &self.u_points,
+//                 &self.v_points,
+//                 elem,
+//                 over_desc_elem,
+//             );
+//             self.computed.insert(desc.clone(), Rc::new(bs));
+//             self.computed.get(&desc).unwrap().clone()
+//         }
+//     }
+// }
+
 /// A utility structure used to generate and cache [BasisFn]'s during Integration
 ///
 /// Three important areas of functionality are handled here:
@@ -54,95 +147,7 @@ pub trait ShapeFn: Clone + Sync + Send + std::fmt::Debug {
 /// * Handle interpolation between parametric spaces to set up inter-layer integration
 /// * Cache computed [BasisFn]'s to avoid re-computation.
 ///
-#[derive(Debug)]
 pub struct BasisFnSampler<SF: ShapeFn> {
-    /// Maximum u-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`
-    pub i_max: usize,
-    /// Maximum v-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`              
-    pub j_max: usize,
-    /// Whether the 2nd derivatives of the [ShapeFn]s will be computed. If true, the endpoints [-1, +1] are also defined along both axes!                 
-    pub compute_d2: bool,
-    /// Gauss Legendre Quadrature points evaluated along u-direction. Defined from (-1 to +1)         
-    u_points: Vec<f64>,
-    /// Gauss Legendre Quadrature points evaluated along v-direction. Defined from (-1 to +1)
-    v_points: Vec<f64>,
-
-    computed: HashMap<BSDescription, Rc<BasisFn<SF>>>,
-}
-
-impl<SF: ShapeFn> BasisFnSampler<SF> {
-    /// Construct a sampler with the following parameters:
-    ///
-    /// * `i_max` : maximum u-directed expansion order
-    /// * `j_max` : maximum v-directed expansion order
-    /// * `num_u_points` : number of Gauss-Leg-Quad points to define along the u-axis. If `None`: a default value is used
-    /// * `num_v_points` : number of Gauss-Leg-Quad points to define along the v-axis. If `None`: a default value is used
-    /// * `compute_2nd_derivs`: whether or not to compute the second derivatives of the [ShapeFn]s.
-    ///
-    /// A [ShapeFn] implementation must also be specified.
-    ///
-    /// These parameters will be used to construct all [BasisFn]s generated by this sampler.
-    ///
-    /// The default values for `num_u_points` and `num_v_points` are cautiously large. Smaller values will result in much faster computation, but may result in a loss of accuracy depending on the implementation of [BasisFn]. For `KOLBasis`, n+2 is sufficient for accurate integration.
-    pub fn with(
-        i_max: usize,
-        j_max: usize,
-        num_u_points: Option<usize>,
-        num_v_points: Option<usize>,
-        compute_2nd_derivs: bool,
-    ) -> (Self, [Vec<f64>; 2]) {
-        let (u_points, u_weights) = gauss_quadrature_points(
-            num_u_points.unwrap_or_else(|| default_ngq(i_max)),
-            compute_2nd_derivs,
-        );
-        let (v_points, v_weights) = gauss_quadrature_points(
-            num_v_points.unwrap_or_else(|| default_ngq(j_max)),
-            compute_2nd_derivs,
-        );
-
-        (
-            Self {
-                i_max,
-                j_max,
-                compute_d2: compute_2nd_derivs,
-                u_points,
-                v_points,
-                computed: HashMap::new(),
-            },
-            [u_weights, v_weights],
-        )
-    }
-
-    /// Generate or retrieve a [BasisFn] defined over an [Elem]. Can be defined over a subset of the Element.
-    ///
-    /// If a descendant `Elem` is provided the [BasisFn] will be sampled over the region of the `Elem` that is overlapped by the descendant. Otherwise, the [BasisFn] will be sampled over the entire `Elem`.
-    pub fn sample_basis_fn(
-        &mut self,
-        elem: &Elem,
-        over_desc_elem: Option<&Elem>,
-    ) -> Rc<BasisFn<SF>> {
-        let desc = BSDescription::new(elem, over_desc_elem);
-
-        if let Some(computed_bs) = self.computed.get(&desc) {
-            computed_bs.clone()
-        } else {
-            let bs = BasisFn::mapped_over_desc(
-                self.i_max,
-                self.j_max,
-                self.compute_d2,
-                &self.u_points,
-                &self.v_points,
-                elem,
-                over_desc_elem,
-            );
-            self.computed.insert(desc.clone(), Rc::new(bs));
-            self.computed.get(&desc).unwrap().clone()
-        }
-    }
-}
-
-/// Same as [BasisFnSampler], except thread-safe internal data structures are used
-pub struct ParBasisFnSampler<SF: ShapeFn> {
     /// Maximum u-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`
     pub i_max: usize,
     /// Maximum v-directed expansion order. [BasisFn]s Generated by this sampler will be defined up to this order in `i`              
@@ -157,7 +162,7 @@ pub struct ParBasisFnSampler<SF: ShapeFn> {
     computed: Arc<Mutex<HashMap<BSDescription, Arc<BasisFn<SF>>>>>,
 }
 
-impl<SF: ShapeFn> ParBasisFnSampler<SF> {
+impl<SF: ShapeFn> BasisFnSampler<SF> {
     /// Construct a thread-safe sampler with the following parameters:
     ///
     /// * `i_max` : maximum u-directed expansion order
@@ -230,6 +235,19 @@ impl<SF: ShapeFn> ParBasisFnSampler<SF> {
                 elem,
                 over_desc_elem,
             )),
+        }
+    }
+}
+
+impl<SF: ShapeFn> Clone for BasisFnSampler<SF> {
+    fn clone(&self) -> Self {
+        Self {
+            i_max: self.i_max,
+            j_max: self.j_max,
+            compute_d2: self.compute_d2,
+            u_points: self.u_points.clone(),
+            v_points: self.v_points.clone(),
+            computed: self.computed.clone(),
         }
     }
 }
