+++ src/fem_domain/domain.rs
@@ -5,19 +5,13 @@ pub mod fields;
 /// The internal geometric structure of a Domain. This is modified by hp-refinements.
 pub mod mesh;
 
-use super::basis::{BasisFnSampler, ParBasisFnSampler, ShapeFn};
-use crate::matrix_math::integration::Integral;
-use crate::matrix_math::linalg::{sparse_matrix::SparseMatrix, GEP};
 use dof::{
     basis_spec::{BSAddress, BasisDir, BasisLoc, BasisSpec},
     DoF,
 };
 use mesh::*;
-
-use rayon::prelude::*;
 use smallvec::smallvec;
 use std::collections::BTreeMap;
-use std::sync::{Arc, Mutex};
 
 /// High Level Description of an FEM Domain
 pub struct Domain {
@@ -336,292 +330,292 @@ impl Domain {
         BSAddress::new(elem_id, elem_idx)
     }
 
-    /// Fill two system matrices using the domains it's Basis Space as the Testing Space. Return a Generalized Eigenproblem ([GEP])
-    ///
-    /// All pairs of overlapping Shape Functions will be integrated and stored in the matrices by their associated DoF IDs
-    ///
-    /// * Two [Integral]s: `AI` and `BI` must be specified. These are used to populate the A and B matrices respectively
-    /// * A [ShapeFn] `SF` must also be specified. This is used to evaluate the Domains [BasisSpec]s
-    pub fn galerkin_sample_gep<SF, AI, BI>(&self, num_gauss_quad: Option<usize>) -> GEP
-    where
-        SF: ShapeFn,
-        AI: Integral,
-        BI: Integral,
-    {
-        // construct an eigenproblem with a and b matrices
-        let mut gep = GEP::new(self.dofs.len());
+    // /// Fill two system matrices using the domains it's Basis Space as the Testing Space. Return a Generalized Eigenproblem ([GEP])
+    // ///
+    // /// All pairs of overlapping Shape Functions will be integrated and stored in the matrices by their associated DoF IDs
+    // ///
+    // /// * Two [Integral]s: `AI` and `BI` must be specified. These are used to populate the A and B matrices respectively
+    // /// * A [ShapeFn] `SF` must also be specified. This is used to evaluate the Domains [BasisSpec]s
+    // pub fn galerkin_sample_gep<SF, AI, BI>(&self, num_gauss_quad: Option<usize>) -> GEP
+    // where
+    //     SF: ShapeFn,
+    //     AI: Integral,
+    //     BI: Integral,
+    // {
+    //     // construct an eigenproblem with a and b matrices
+    //     let mut gep = GEP::new(self.dofs.len());
 
-        // construct basis sampler
-        let [i_max, j_max] = self.mesh.max_expansion_orders();
-        let (mut bf_sampler, [u_weights, v_weights]): (BasisFnSampler<SF>, _) =
-            BasisFnSampler::with(
-                i_max as usize,
-                j_max as usize,
-                num_gauss_quad,
-                num_gauss_quad,
-                false,
-            );
+    //     // construct basis sampler
+    //     let [i_max, j_max] = self.mesh.max_expansion_orders();
+    //     let (mut bf_sampler, [u_weights, v_weights]): (BasisFnSampler<SF>, _) =
+    //         BasisFnSampler::with(
+    //             i_max as usize,
+    //             j_max as usize,
+    //             num_gauss_quad,
+    //             num_gauss_quad,
+    //             false,
+    //         );
 
-        // setup integration
-        let a_integrator = AI::with_weights(&u_weights, &v_weights);
-        let b_integrator = BI::with_weights(&u_weights, &v_weights);
+    //     // setup integration
+    //     let a_integrator = AI::with_weights(&u_weights, &v_weights);
+    //     let b_integrator = BI::with_weights(&u_weights, &v_weights);
 
-        for elem in self.elems() {
-            // get relevant data for this Elem
-            let local_basis_specs = self.local_basis_specs(elem.id).unwrap();
-            let desc_basis_specs = self.descendant_basis_specs(elem.id).unwrap();
+    //     for elem in self.elems() {
+    //         // get relevant data for this Elem
+    //         let local_basis_specs = self.local_basis_specs(elem.id).unwrap();
+    //         let desc_basis_specs = self.descendant_basis_specs(elem.id).unwrap();
 
-            let bs_local = bf_sampler.sample_basis_fn(elem, None);
-            let elem_materials = elem.get_materials();
+    //         let bs_local = bf_sampler.sample_basis_fn(elem, None);
+    //         let elem_materials = elem.get_materials();
 
-            let mut local_a_entries: Vec<([usize; 2], f64)> =
-                Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
-            let mut local_b_entries: Vec<([usize; 2], f64)> =
-                Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
+    //         let mut local_a_entries: Vec<([usize; 2], f64)> =
+    //             Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
+    //         let mut local_b_entries: Vec<([usize; 2], f64)> =
+    //             Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
 
-            // local - local
-            for (i, (p_orders, p_dir, p_dof_id)) in local_basis_specs
-                .iter()
-                .map(|bs_p| bs_p.integration_data())
-                .enumerate()
-            {
-                for (q_orders, q_dir, q_dof_id) in local_basis_specs
-                    .iter()
-                    .skip(i)
-                    .map(|bs_q| bs_q.integration_data())
-                {
-                    let a = a_integrator
-                        .integrate(
-                            p_dir,
-                            q_dir,
-                            p_orders,
-                            q_orders,
-                            &bs_local,
-                            &bs_local,
-                            &elem_materials,
-                        )
-                        .full_solution();
-                    let b = b_integrator
-                        .integrate(
-                            p_dir,
-                            q_dir,
-                            p_orders,
-                            q_orders,
-                            &bs_local,
-                            &bs_local,
-                            &elem_materials,
-                        )
-                        .full_solution();
+    //         // local - local
+    //         for (i, (p_orders, p_dir, p_dof_id)) in local_basis_specs
+    //             .iter()
+    //             .map(|bs_p| bs_p.integration_data())
+    //             .enumerate()
+    //         {
+    //             for (q_orders, q_dir, q_dof_id) in local_basis_specs
+    //                 .iter()
+    //                 .skip(i)
+    //                 .map(|bs_q| bs_q.integration_data())
+    //             {
+    //                 let a = a_integrator
+    //                     .integrate(
+    //                         p_dir,
+    //                         q_dir,
+    //                         p_orders,
+    //                         q_orders,
+    //                         &bs_local,
+    //                         &bs_local,
+    //                         &elem_materials,
+    //                     )
+    //                     .full_solution();
+    //                 let b = b_integrator
+    //                     .integrate(
+    //                         p_dir,
+    //                         q_dir,
+    //                         p_orders,
+    //                         q_orders,
+    //                         &bs_local,
+    //                         &bs_local,
+    //                         &elem_materials,
+    //                     )
+    //                     .full_solution();
 
-                    local_a_entries.push(([p_dof_id, q_dof_id], a));
-                    local_b_entries.push(([p_dof_id, q_dof_id], b));
-                }
-            }
+    //                 local_a_entries.push(([p_dof_id, q_dof_id], a));
+    //                 local_b_entries.push(([p_dof_id, q_dof_id], b));
+    //             }
+    //         }
 
-            gep.a.insert_group(local_a_entries);
-            gep.b.insert_group(local_b_entries);
+    //         gep.a.insert_group(local_a_entries);
+    //         gep.b.insert_group(local_b_entries);
 
-            let mut desc_a_entries: Vec<([usize; 2], f64)> =
-                Vec::with_capacity(local_basis_specs.len() * desc_basis_specs.len());
-            let mut desc_b_entries: Vec<([usize; 2], f64)> =
-                Vec::with_capacity(local_basis_specs.len() * desc_basis_specs.len());
+    //         let mut desc_a_entries: Vec<([usize; 2], f64)> =
+    //             Vec::with_capacity(local_basis_specs.len() * desc_basis_specs.len());
+    //         let mut desc_b_entries: Vec<([usize; 2], f64)> =
+    //             Vec::with_capacity(local_basis_specs.len() * desc_basis_specs.len());
 
-            // local - desc
-            for (p_orders, p_dir, p_dof_id) in
-                local_basis_specs.iter().map(|bs_p| bs_p.integration_data())
-            {
-                for &(q_elem_id, q_elem_basis_specs) in desc_basis_specs.iter() {
-                    let bs_p_sampled =
-                        bf_sampler.sample_basis_fn(elem, Some(&self.mesh.elems[q_elem_id]));
-                    let bs_q_local = bf_sampler.sample_basis_fn(&self.mesh.elems[q_elem_id], None);
+    //         // local - desc
+    //         for (p_orders, p_dir, p_dof_id) in
+    //             local_basis_specs.iter().map(|bs_p| bs_p.integration_data())
+    //         {
+    //             for &(q_elem_id, q_elem_basis_specs) in desc_basis_specs.iter() {
+    //                 let bs_p_sampled =
+    //                     bf_sampler.sample_basis_fn(elem, Some(&self.mesh.elems[q_elem_id]));
+    //                 let bs_q_local = bf_sampler.sample_basis_fn(&self.mesh.elems[q_elem_id], None);
 
-                    for (q_orders, q_dir, q_dof_id) in q_elem_basis_specs
-                        .iter()
-                        .map(|bs_q| bs_q.integration_data())
-                    {
-                        let a = a_integrator
-                            .integrate(
-                                p_dir,
-                                q_dir,
-                                p_orders,
-                                q_orders,
-                                &bs_p_sampled,
-                                &bs_q_local,
-                                &elem_materials,
-                            )
-                            .full_solution();
+    //                 for (q_orders, q_dir, q_dof_id) in q_elem_basis_specs
+    //                     .iter()
+    //                     .map(|bs_q| bs_q.integration_data())
+    //                 {
+    //                     let a = a_integrator
+    //                         .integrate(
+    //                             p_dir,
+    //                             q_dir,
+    //                             p_orders,
+    //                             q_orders,
+    //                             &bs_p_sampled,
+    //                             &bs_q_local,
+    //                             &elem_materials,
+    //                         )
+    //                         .full_solution();
 
-                        let b = b_integrator
-                            .integrate(
-                                p_dir,
-                                q_dir,
-                                p_orders,
-                                q_orders,
-                                &bs_p_sampled,
-                                &bs_q_local,
-                                &elem_materials,
-                            )
-                            .full_solution();
+    //                     let b = b_integrator
+    //                         .integrate(
+    //                             p_dir,
+    //                             q_dir,
+    //                             p_orders,
+    //                             q_orders,
+    //                             &bs_p_sampled,
+    //                             &bs_q_local,
+    //                             &elem_materials,
+    //                         )
+    //                         .full_solution();
 
-                        desc_a_entries.push(([p_dof_id, q_dof_id], a));
-                        desc_b_entries.push(([p_dof_id, q_dof_id], b));
-                    }
-                }
-            }
+    //                     desc_a_entries.push(([p_dof_id, q_dof_id], a));
+    //                     desc_b_entries.push(([p_dof_id, q_dof_id], b));
+    //                 }
+    //             }
+    //         }
 
-            gep.a.insert_group(desc_a_entries);
-            gep.b.insert_group(desc_b_entries);
-        }
+    //         gep.a.insert_group(desc_a_entries);
+    //         gep.b.insert_group(desc_b_entries);
+    //     }
 
-        gep
-    }
+    //     gep
+    // }
 
-    /// Same as `galerkin_sample_gep`, except matrices are filled in parallel using the Rayon Global ThreadPool
-    pub fn galerkin_sample_gep_parallel<SF, AI, BI>(&self, num_gauss_quad: Option<usize>) -> GEP
-    where
-        SF: ShapeFn,
-        AI: Integral,
-        BI: Integral,
-    {
-        // construct an eigenproblem with a and b matrices
-        let mut gep = GEP::new(self.dofs.len());
+    // /// Same as `galerkin_sample_gep`, except matrices are filled in parallel using the Rayon Global ThreadPool
+    // pub fn galerkin_sample_gep_parallel<SF, AI, BI>(&self, num_gauss_quad: Option<usize>) -> GEP
+    // where
+    //     SF: ShapeFn,
+    //     AI: Integral,
+    //     BI: Integral,
+    // {
+    //     // construct an eigenproblem with a and b matrices
+    //     let mut gep = GEP::new(self.dofs.len());
 
-        // construct basis sampler
-        let [i_max, j_max] = self.mesh.max_expansion_orders();
-        let (bs_sampler, [u_weights, v_weights]): (ParBasisFnSampler<SF>, _) =
-            ParBasisFnSampler::with(
-                i_max as usize,
-                j_max as usize,
-                num_gauss_quad,
-                num_gauss_quad,
-                false,
-            );
-        let bf_sampler_send = Arc::new(Mutex::new(bs_sampler));
+    //     // construct basis sampler
+    //     let [i_max, j_max] = self.mesh.max_expansion_orders();
+    //     let (bs_sampler, [u_weights, v_weights]): (ParBasisFnSampler<SF>, _) =
+    //         ParBasisFnSampler::with(
+    //             i_max as usize,
+    //             j_max as usize,
+    //             num_gauss_quad,
+    //             num_gauss_quad,
+    //             false,
+    //         );
+    //     let bf_sampler_send = Arc::new(Mutex::new(bs_sampler));
 
-        // setup integration
-        let a_integrator = AI::with_weights(&u_weights, &v_weights);
-        let b_integrator = BI::with_weights(&u_weights, &v_weights);
+    //     // setup integration
+    //     let a_integrator = AI::with_weights(&u_weights, &v_weights);
+    //     let b_integrator = BI::with_weights(&u_weights, &v_weights);
 
-        gep.par_extend(self.mesh.elems.par_iter().map(|elem| {
-            let mut local_a = SparseMatrix::new(self.dofs.len());
-            let mut local_b = SparseMatrix::new(self.dofs.len());
+    //     gep.par_extend(self.mesh.elems.par_iter().map(|elem| {
+    //         let mut local_a = SparseMatrix::new(self.dofs.len());
+    //         let mut local_b = SparseMatrix::new(self.dofs.len());
 
-            let bf_sampler_elem = bf_sampler_send.clone();
+    //         let bf_sampler_elem = bf_sampler_send.clone();
 
-            let elem_materials = elem.get_materials();
+    //         let elem_materials = elem.get_materials();
 
-            // get relevant data for this Elem
-            let bs_local = bf_sampler_elem.lock().unwrap().sample_basis_fn(elem, None);
-            let local_basis_specs = self.local_basis_specs(elem.id).unwrap();
-            let desc_basis_specs = self.descendant_basis_specs(elem.id).unwrap();
+    //         // get relevant data for this Elem
+    //         let bs_local = bf_sampler_elem.lock().unwrap().sample_basis_fn(elem, None);
+    //         let local_basis_specs = self.local_basis_specs(elem.id).unwrap();
+    //         let desc_basis_specs = self.descendant_basis_specs(elem.id).unwrap();
 
-            let mut local_a_entries: Vec<([usize; 2], f64)> =
-                Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
-            let mut local_b_entries: Vec<([usize; 2], f64)> =
-                Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
+    //         let mut local_a_entries: Vec<([usize; 2], f64)> =
+    //             Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
+    //         let mut local_b_entries: Vec<([usize; 2], f64)> =
+    //             Vec::with_capacity(local_basis_specs.len() * local_basis_specs.len() / 2);
 
-            // local - local
-            for (i, (p_orders, p_dir, p_dof_id)) in local_basis_specs
-                .iter()
-                .map(|bs_p| bs_p.integration_data())
-                .enumerate()
-            {
-                for (q_orders, q_dir, q_dof_id) in local_basis_specs
-                    .iter()
-                    .skip(i)
-                    .map(|bs_q| bs_q.integration_data())
-                {
-                    let a = a_integrator
-                        .integrate(
-                            p_dir,
-                            q_dir,
-                            p_orders,
-                            q_orders,
-                            &bs_local,
-                            &bs_local,
-                            elem_materials,
-                        )
-                        .full_solution();
-                    let b = b_integrator
-                        .integrate(
-                            p_dir,
-                            q_dir,
-                            p_orders,
-                            q_orders,
-                            &bs_local,
-                            &bs_local,
-                            elem_materials,
-                        )
-                        .full_solution();
+    //         // local - local
+    //         for (i, (p_orders, p_dir, p_dof_id)) in local_basis_specs
+    //             .iter()
+    //             .map(|bs_p| bs_p.integration_data())
+    //             .enumerate()
+    //         {
+    //             for (q_orders, q_dir, q_dof_id) in local_basis_specs
+    //                 .iter()
+    //                 .skip(i)
+    //                 .map(|bs_q| bs_q.integration_data())
+    //             {
+    //                 let a = a_integrator
+    //                     .integrate(
+    //                         p_dir,
+    //                         q_dir,
+    //                         p_orders,
+    //                         q_orders,
+    //                         &bs_local,
+    //                         &bs_local,
+    //                         elem_materials,
+    //                     )
+    //                     .full_solution();
+    //                 let b = b_integrator
+    //                     .integrate(
+    //                         p_dir,
+    //                         q_dir,
+    //                         p_orders,
+    //                         q_orders,
+    //                         &bs_local,
+    //                         &bs_local,
+    //                         elem_materials,
+    //                     )
+    //                     .full_solution();
 
-                    local_a_entries.push(([p_dof_id, q_dof_id], a));
-                    local_b_entries.push(([p_dof_id, q_dof_id], b));
-                }
-            }
+    //                 local_a_entries.push(([p_dof_id, q_dof_id], a));
+    //                 local_b_entries.push(([p_dof_id, q_dof_id], b));
+    //             }
+    //         }
 
-            local_a.insert_group(local_a_entries);
-            local_b.insert_group(local_b_entries);
+    //         local_a.insert_group(local_a_entries);
+    //         local_b.insert_group(local_b_entries);
 
-            let mut desc_a_entries: Vec<([usize; 2], f64)> =
-                Vec::with_capacity(local_basis_specs.len() * desc_basis_specs.len());
-            let mut desc_b_entries: Vec<([usize; 2], f64)> =
-                Vec::with_capacity(local_basis_specs.len() * desc_basis_specs.len());
+    //         let mut desc_a_entries: Vec<([usize; 2], f64)> =
+    //             Vec::with_capacity(local_basis_specs.len() * desc_basis_specs.len());
+    //         let mut desc_b_entries: Vec<([usize; 2], f64)> =
+    //             Vec::with_capacity(local_basis_specs.len() * desc_basis_specs.len());
 
-            // local - desc
-            for (p_orders, p_dir, p_dof_id) in
-                local_basis_specs.iter().map(|bs_p| bs_p.integration_data())
-            {
-                for &(q_elem_id, q_elem_basis_specs) in desc_basis_specs.iter() {
-                    let bs_p_sampled = bf_sampler_elem
-                        .lock()
-                        .unwrap()
-                        .sample_basis_fn(elem, Some(&self.mesh.elems[q_elem_id]));
-                    let bs_q_local = bf_sampler_elem
-                        .lock()
-                        .unwrap()
-                        .sample_basis_fn(&self.mesh.elems[q_elem_id], None);
+    //         // local - desc
+    //         for (p_orders, p_dir, p_dof_id) in
+    //             local_basis_specs.iter().map(|bs_p| bs_p.integration_data())
+    //         {
+    //             for &(q_elem_id, q_elem_basis_specs) in desc_basis_specs.iter() {
+    //                 let bs_p_sampled = bf_sampler_elem
+    //                     .lock()
+    //                     .unwrap()
+    //                     .sample_basis_fn(elem, Some(&self.mesh.elems[q_elem_id]));
+    //                 let bs_q_local = bf_sampler_elem
+    //                     .lock()
+    //                     .unwrap()
+    //                     .sample_basis_fn(&self.mesh.elems[q_elem_id], None);
 
-                    for (q_orders, q_dir, q_dof_id) in q_elem_basis_specs
-                        .iter()
-                        .map(|bs_q| bs_q.integration_data())
-                    {
-                        let a = a_integrator
-                            .integrate(
-                                p_dir,
-                                q_dir,
-                                p_orders,
-                                q_orders,
-                                &bs_p_sampled,
-                                &bs_q_local,
-                                elem_materials,
-                            )
-                            .full_solution();
-                        let b = b_integrator
-                            .integrate(
-                                p_dir,
-                                q_dir,
-                                p_orders,
-                                q_orders,
-                                &bs_p_sampled,
-                                &bs_q_local,
-                                elem_materials,
-                            )
-                            .full_solution();
+    //                 for (q_orders, q_dir, q_dof_id) in q_elem_basis_specs
+    //                     .iter()
+    //                     .map(|bs_q| bs_q.integration_data())
+    //                 {
+    //                     let a = a_integrator
+    //                         .integrate(
+    //                             p_dir,
+    //                             q_dir,
+    //                             p_orders,
+    //                             q_orders,
+    //                             &bs_p_sampled,
+    //                             &bs_q_local,
+    //                             elem_materials,
+    //                         )
+    //                         .full_solution();
+    //                     let b = b_integrator
+    //                         .integrate(
+    //                             p_dir,
+    //                             q_dir,
+    //                             p_orders,
+    //                             q_orders,
+    //                             &bs_p_sampled,
+    //                             &bs_q_local,
+    //                             elem_materials,
+    //                         )
+    //                         .full_solution();
 
-                        desc_a_entries.push(([p_dof_id, q_dof_id], a));
-                        desc_b_entries.push(([p_dof_id, q_dof_id], b));
-                    }
-                }
-            }
+    //                     desc_a_entries.push(([p_dof_id, q_dof_id], a));
+    //                     desc_b_entries.push(([p_dof_id, q_dof_id], b));
+    //                 }
+    //             }
+    //         }
 
-            local_a.insert_group(desc_a_entries);
-            local_b.insert_group(desc_b_entries);
+    //         local_a.insert_group(desc_a_entries);
+    //         local_b.insert_group(desc_b_entries);
 
-            [local_a, local_b]
-        }));
+    //         [local_a, local_b]
+    //     }));
 
-        gep
-    }
+    //     gep
+    // }
 }
 
 struct IdTracker {
